Module aws_sdk_eventbridge::error

Modules

concurrent_modification_exception

See ConcurrentModificationException

illegal_status_exception

See IllegalStatusException

internal_exception

See InternalException

invalid_event_pattern_exception

See InvalidEventPatternException

invalid_state_exception

See InvalidStateException

limit_exceeded_exception

See LimitExceededException

managed_rule_exception

See ManagedRuleException

operation_disabled_exception

See OperationDisabledException

policy_length_exceeded_exception

See PolicyLengthExceededException

resource_already_exists_exception

See ResourceAlreadyExistsException

resource_not_found_exception

See ResourceNotFoundException

Structs

ActivateEventSourceError

CancelReplayError

ConcurrentModificationException

There is concurrent modification on a rule, target, archive, or replay.

CreateApiDestinationError

CreateArchiveError

CreateConnectionError

CreateEventBusError

CreatePartnerEventSourceError

DeactivateEventSourceError

DeauthorizeConnectionError

DeleteApiDestinationError

DeleteArchiveError

DeleteConnectionError

DeleteEventBusError

DeletePartnerEventSourceError

DeleteRuleError

DescribeApiDestinationError

DescribeArchiveError

DescribeConnectionError

DescribeEventBusError

DescribeEventSourceError

DescribePartnerEventSourceError

DescribeReplayError

DescribeRuleError

DisableRuleError

EnableRuleError

IllegalStatusException

An error occurred because a replay can be canceled only when the state is Running or

InternalException

This exception occurs due to unexpected causes.

InvalidEventPatternException

The event pattern is not valid.

InvalidStateException

The specified state is not a valid state for an event source.

LimitExceededException

The request failed because it attempted to create resource beyond the allowed service

ListApiDestinationsError

ListArchivesError

ListConnectionsError

ListEventBusesError

ListEventSourcesError

ListPartnerEventSourceAccountsError

ListPartnerEventSourcesError

ListReplaysError

ListRuleNamesByTargetError

ListRulesError

ListTagsForResourceError

ListTargetsByRuleError

ManagedRuleException

This rule was created by an Amazon Web Services service on behalf of your account. It is managed by that

OperationDisabledException

The operation you are attempting is not available in this region.

PolicyLengthExceededException

The event bus policy is too long. For more information, see the limits.

PutEventsError

PutPartnerEventsError

PutPermissionError

PutRuleError

PutTargetsError

RemovePermissionError

RemoveTargetsError

ResourceAlreadyExistsException

The resource you are trying to create already exists.

ResourceNotFoundException

An entity that you specified does not exist.

StartReplayError

TagResourceError

TestEventPatternError

UntagResourceError

UpdateApiDestinationError

UpdateArchiveError

UpdateConnectionError

Enums

ActivateEventSourceErrorKind

CancelReplayErrorKind

CreateApiDestinationErrorKind

CreateArchiveErrorKind

CreateConnectionErrorKind

CreateEventBusErrorKind

CreatePartnerEventSourceErrorKind

DeactivateEventSourceErrorKind

DeauthorizeConnectionErrorKind

DeleteApiDestinationErrorKind

DeleteArchiveErrorKind

DeleteConnectionErrorKind

DeleteEventBusErrorKind

DeletePartnerEventSourceErrorKind

DeleteRuleErrorKind

DescribeApiDestinationErrorKind

DescribeArchiveErrorKind

DescribeConnectionErrorKind

DescribeEventBusErrorKind

DescribeEventSourceErrorKind

DescribePartnerEventSourceErrorKind

DescribeReplayErrorKind

DescribeRuleErrorKind

DisableRuleErrorKind

EnableRuleErrorKind

ListApiDestinationsErrorKind

ListArchivesErrorKind

ListConnectionsErrorKind

ListEventBusesErrorKind

ListEventSourcesErrorKind

ListPartnerEventSourceAccountsErrorKind

ListPartnerEventSourcesErrorKind

ListReplaysErrorKind

ListRuleNamesByTargetErrorKind

ListRulesErrorKind

ListTagsForResourceErrorKind

ListTargetsByRuleErrorKind

PutEventsErrorKind

PutPartnerEventsErrorKind

PutPermissionErrorKind

PutRuleErrorKind

PutTargetsErrorKind

RemovePermissionErrorKind

RemoveTargetsErrorKind

StartReplayErrorKind

TagResourceErrorKind

TestEventPatternErrorKind

UntagResourceErrorKind

UpdateApiDestinationErrorKind

UpdateArchiveErrorKind

UpdateConnectionErrorKind