Struct aws_sdk_eventbridge::model::ecs_parameters::Builder

#[non_exhaustive]
pub struct Builder { /* fields omitted */ }

A builder for EcsParameters

Implementations

impl Builder

pub fn task_definition_arn(self, input: impl Into<String>) -> Self

The ARN of the task definition to use if the event target is an Amazon ECS task.

pub fn set_task_definition_arn(self, input: Option<String>) -> Self

pub fn task_count(self, input: i32) -> Self

The number of tasks to create based on TaskDefinition. The default is 1.

pub fn set_task_count(self, input: Option<i32>) -> Self

pub fn launch_type(self, input: LaunchType) -> Self

Specifies the launch type on which your task is running. The launch type that you specify here must match one of the launch type (compatibilities) of the target task. The FARGATE value is supported only in the Regions where Fargate witt Amazon ECS is supported. For more information, see Fargate on Amazon ECS in the Amazon Elastic Container Service Developer Guide.

pub fn set_launch_type(self, input: Option<LaunchType>) -> Self

pub fn network_configuration(self, input: NetworkConfiguration) -> Self

Use this structure if the Amazon ECS task uses the awsvpc network mode. This structure specifies the VPC subnets and security groups associated with the task, and whether a public IP address is to be used. This structure is required if LaunchType is FARGATE because the awsvpc mode is required for Fargate tasks.

If you specify NetworkConfiguration when the target ECS task does not use the awsvpc network mode, the task fails.

pub fn set_network_configuration(
self,
input: Option<NetworkConfiguration>
) -> Self

pub fn platform_version(self, input: impl Into<String>) -> Self

Specifies the platform version for the task. Specify only the numeric portion of the platform version, such as 1.1.0.

This structure is used only if LaunchType is FARGATE. For more information about valid platform versions, see Fargate Platform Versions in the Amazon Elastic Container Service Developer Guide.

pub fn set_platform_version(self, input: Option<String>) -> Self

pub fn group(self, input: impl Into<String>) -> Self

Specifies an ECS task group for the task. The maximum length is 255 characters.

pub fn set_group(self, input: Option<String>) -> Self

pub fn capacity_provider_strategy(
self,
input: impl Into<CapacityProviderStrategyItem>
) -> Self

pub fn set_capacity_provider_strategy(
self,
input: Option<Vec<CapacityProviderStrategyItem>>
) -> Self

pub fn enable_ecs_managed_tags(self, input: bool) -> Self

Specifies whether to enable Amazon ECS managed tags for the task. For more information, see Tagging Your Amazon ECS Resources in the Amazon Elastic Container Service Developer Guide.

pub fn set_enable_ecs_managed_tags(self, input: Option<bool>) -> Self

pub fn enable_execute_command(self, input: bool) -> Self

Whether or not to enable the execute command functionality for the containers in this task. If true, this enables execute command functionality on all containers in the task.

pub fn set_enable_execute_command(self, input: Option<bool>) -> Self

pub fn placement_constraints(
self,
input: impl Into<PlacementConstraint>
) -> Self

pub fn set_placement_constraints(
self,
input: Option<Vec<PlacementConstraint>>
) -> Self

pub fn placement_strategy(self, input: impl Into<PlacementStrategy>) -> Self

pub fn set_placement_strategy(
self,
input: Option<Vec<PlacementStrategy>>
) -> Self

pub fn propagate_tags(self, input: PropagateTags) -> Self

Specifies whether to propagate the tags from the task definition to the task. If no value is specified, the tags are not propagated. Tags can only be propagated to the task during task creation. To add tags to a task after task creation, use the TagResource API action.

pub fn set_propagate_tags(self, input: Option<PropagateTags>) -> Self

pub fn reference_id(self, input: impl Into<String>) -> Self

The reference ID to use for the task.

pub fn set_reference_id(self, input: Option<String>) -> Self

pub fn tags(self, input: impl Into<Tag>) -> Self

pub fn set_tags(self, input: Option<Vec<Tag>>) -> Self

pub fn build(self) -> EcsParameters

Consumes the builder and constructs a EcsParameters