Struct aws_sdk_eventbridge::model::Target

#[non_exhaustive]
pub struct Target {

    pub id: Option<String>,
    pub arn: Option<String>,
    pub role_arn: Option<String>,
    pub input: Option<String>,
    pub input_path: Option<String>,
    pub input_transformer: Option<InputTransformer>,
    pub kinesis_parameters: Option<KinesisParameters>,
    pub run_command_parameters: Option<RunCommandParameters>,
    pub ecs_parameters: Option<EcsParameters>,
    pub batch_parameters: Option<BatchParameters>,
    pub sqs_parameters: Option<SqsParameters>,
    pub http_parameters: Option<HttpParameters>,
    pub redshift_data_parameters: Option<RedshiftDataParameters>,
    pub sage_maker_pipeline_parameters: Option<SageMakerPipelineParameters>,
    pub dead_letter_config: Option<DeadLetterConfig>,
    pub retry_policy: Option<RetryPolicy>,

}

Targets are the resources to be invoked when a rule is triggered. For a complete list of services and resources that can be set as a target, see PutTargets.

If you are setting the event bus of another account as the target, and that account granted permission to your account through an organization instead of directly by the account ID, then you must specify a RoleArn with proper permissions in the Target structure. For more information, see Sending and Receiving Events Between Amazon Web Services Accounts in the Amazon EventBridge User Guide.

Fields (Non-exhaustive)

Non-exhaustive structs could have additional fields added in future. Therefore, non-exhaustive structs cannot be constructed in external crates using the traditional Struct { .. } syntax; cannot be matched against without a wildcard ..; and struct update syntax will not work.

id: Option<String>

The ID of the target. We recommend using a memorable and unique string.

arn: Option<String>

The Amazon Resource Name (ARN) of the target.

role_arn: Option<String>

The Amazon Resource Name (ARN) of the IAM role to be used for this target when the rule is triggered. If one rule triggers multiple targets, you can use a different IAM role for each target.

input: Option<String>

Valid JSON text passed to the target. In this case, nothing from the event itself is passed to the target. For more information, see The JavaScript Object Notation (JSON) Data Interchange Format.

input_path: Option<String>

The value of the JSONPath that is used for extracting part of the matched event when passing it to the target. You must use JSON dot notation, not bracket notation. For more information about JSON paths, see JSONPath.

input_transformer: Option<InputTransformer>

Settings to enable you to provide custom input to a target based on certain event data. You can extract one or more key-value pairs from the event and then use that data to send customized input to the target.

kinesis_parameters: Option<KinesisParameters>

The custom parameter you can use to control the shard assignment, when the target is a Kinesis data stream. If you do not include this parameter, the default is to use the eventId as the partition key.

run_command_parameters: Option<RunCommandParameters>

Parameters used when you are using the rule to invoke Amazon EC2 Run Command.

ecs_parameters: Option<EcsParameters>

Contains the Amazon ECS task definition and task count to be used, if the event target is an Amazon ECS task. For more information about Amazon ECS tasks, see Task Definitions in the Amazon EC2 Container Service Developer Guide.

batch_parameters: Option<BatchParameters>

If the event target is an Batch job, this contains the job definition, job name, and other parameters. For more information, see Jobs in the Batch User Guide.

sqs_parameters: Option<SqsParameters>

Contains the message group ID to use when the target is a FIFO queue.

If you specify an SQS FIFO queue as a target, the queue must have content-based deduplication enabled.

http_parameters: Option<HttpParameters>

Contains the HTTP parameters to use when the target is a API Gateway REST endpoint or EventBridge ApiDestination.

If you specify an API Gateway REST API or EventBridge ApiDestination as a target, you can use this parameter to specify headers, path parameters, and query string keys/values as part of your target invoking request. If you're using ApiDestinations, the corresponding Connection can also have these values configured. In case of any conflicting keys, values from the Connection take precedence.

redshift_data_parameters: Option<RedshiftDataParameters>

Contains the Amazon Redshift Data API parameters to use when the target is a Amazon Redshift cluster.

If you specify a Amazon Redshift Cluster as a Target, you can use this to specify parameters to invoke the Amazon Redshift Data API ExecuteStatement based on EventBridge events.

sage_maker_pipeline_parameters: Option<SageMakerPipelineParameters>

Contains the SageMaker Model Building Pipeline parameters to start execution of a SageMaker Model Building Pipeline.

If you specify a SageMaker Model Building Pipeline as a target, you can use this to specify parameters to start a pipeline execution based on EventBridge events.

dead_letter_config: Option<DeadLetterConfig>

The DeadLetterConfig that defines the target queue to send dead-letter queue events to.

retry_policy: Option<RetryPolicy>

The RetryPolicy object that contains the retry policy configuration to use for the dead-letter queue.

Implementations

impl Target

pub fn builder() -> Builder

Creates a new builder-style object to manufacture Target